requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.1.7
google-generativeai==0.8.4
pydantic==2.9.2
pydantic-settings==2.6.0
python-multipart==0.0.12
